refactor: menu, validator 패키지 추가 및 파일 이동
