+++ src/main/java/subway/view/validator/MenuSelectionException.java
@@ -1,4 +1,4 @@
-package subway.view;
+package subway.view.validator;
 
 public class MenuSelectionException extends RuntimeException {
     public MenuSelectionException(String message) {
